Create CommonCard Component for ListCard and MapCard Components.
